fix(next): removes outputPath from project json dev target build if the same as sourceRoot
The configuration is not happy when the src and destination of a build bundle is being generated. If
no outputPath is supplied, the bundle is automatically created inside the dist folder which is fine.
Hence the fix is just a bit of cleanup whereby removing the config if the supplied path is the same
as project source.